Transactions: Don't load save transaction items at any time
We were loading the saved transaction item data to show in the form, no
matter if it was from a partially completed transaction form, or a previous
instance.
Now we load the transaction data if the user came from a transaction create
route or if the requester doesn't specify otherwise.