Transactions: Load correct partial data on form error
If there were problems with the transaction items for example, we were
trying to load the items stored in the session data, instead of loading
the old values.
From now on, if we got any errors with the for during loading, we will do
an 'emergency' load from the old() values.