Transactions: Fix icon not showing up when adding multiple items
If the user was adding more than one item field to the transaction at
the same time, in many cases the trash icon wouldn't show up, just for
the first item. Found a workaround that strangely works. If the icons
don't show up this way, hopefully increasing the timeout value a bit
will help.